after make BackEnd Permissions

+++ app/Http/Controllers/API/invoiceApi.php
@@ -18,16 +18,13 @@ class invoiceApi extends Controller
 {
     public function dailyinvoice(Request $request)
     {
-        $store_id = $request->store_id;
         $this->validate($request, [
             'getto' => 'required',
             'getfrom' => 'required',
             'store_id' => 'required|integer',
         ]);
-        // $getToDate = date('Y-m-d', Time() - (60 * 60 * 24 * $request->getto));
-        // $getfromdate = date('Y-m-d', Time() - (60 * 60 * 24 * $request->getfrom));
-        $store = store::find($store_id);
-
+        $store_id = $request->store_id;
+        $store = store::find($request->store_id);
         $invoices = invoice::where('store_id', $store_id)->where('paid', '!=', NULL)->where('created_at', '>=', $request->getfrom)->where('created_at', '<=', $request->getto)->with('invoicedets')->orderby('created_at', 'DESC')->get();
         $invoiceTotal = 0;
         foreach ($invoices as $invoice) {
@@ -58,34 +55,48 @@ class invoiceApi extends Controller
         $this->validate($request, [
             'store_id'      => 'required'
         ]);
-        $invoice_s = invoicesett::where('store_id', $request->store_id)->first();
-        $invoice_s->tax = $request->tax;
-        $invoice_s->autotax     = $request->autotax;
-        $invoice_s->tax_record  = $request->tax_record;
-        $invoice_s->tax_card    = $request->tax_card;
-        $invoice_s->file_no     = $request->file_no;
-        $invoice_s->product_rtn = $request->product_rtn;
-        $invoice_s->message_en = $request->message_en;
-        $invoice_s->message_ar = $request->message_ar;
-        $invoice_s->save();
 
-        $historyApi = new historyApi;
-        $des_ar = " تم تعديل اعدادات الفواتير الخاصة بالمتجر";
-        $des_en = " Store billing settings have been modified";
-        $history = $historyApi->createHistory($des_ar, $des_en, $request->store_id, Auth::id());
-        return 'done';
+        $positionApi = new positionApi();
+        $check = $positionApi->checkPositionRoute($request->store_id, Auth::id(), 'invoice_edit');
+        if ($check) {
+            $invoice_s = invoicesett::where('store_id', $request->store_id)->first();
+            $invoice_s->tax = $request->tax;
+            $invoice_s->autotax     = $request->autotax;
+            $invoice_s->tax_record  = $request->tax_record;
+            $invoice_s->tax_card    = $request->tax_card;
+            $invoice_s->file_no     = $request->file_no;
+            $invoice_s->product_rtn = $request->product_rtn;
+            $invoice_s->message_en = $request->message_en;
+            $invoice_s->message_ar = $request->message_ar;
+            $invoice_s->save();
+
+            $historyApi = new historyApi;
+            $des_ar = " تم تعديل اعدادات الفواتير الخاصة بالمتجر";
+            $des_en = " Store billing settings have been modified";
+            $history = $historyApi->createHistory($des_ar, $des_en, $request->store_id, Auth::id());
+            return 'done';
+        } else {
+            return abort(401);
+        }
     }
 
     public function deleteinvoice(Request $request)
     {
-        $invoice = invoice::find($request->invoice_id);
-        $store = store::find($invoice->store_id);
-        $historyApi = new historyApi;
-        $des_ar = " تم حذف الفاتورة رقم $invoice->id ";
-        $des_en = " Invoice No. $invoice->id has been deleted ";
-        $history = $historyApi->createHistory($des_ar, $des_en, $store->id, Auth::id());
-        $invoice->delete();
-        return "Deleted successfully";
+
+        $positionApi = new positionApi();
+        $check = $positionApi->checkPositionRoute($request->store_id, Auth::id(), 'invoice_delete');
+        if ($check) {
+            $invoice = invoice::find($request->invoice_id);
+            $store = store::find($invoice->store_id);
+            $historyApi = new historyApi;
+            $des_ar = " تم حذف الفاتورة رقم $invoice->id ";
+            $des_en = " Invoice No. $invoice->id has been deleted ";
+            $history = $historyApi->createHistory($des_ar, $des_en, $store->id, Auth::id());
+            $invoice->delete();
+            return "Deleted successfully";
+        } else {
+            return abort(401);
+        }
     }
 
     public function payInvoice(Request $request)
@@ -95,21 +106,28 @@ class invoiceApi extends Controller
             'table_id' => 'required|numeric',
             'paidamount' => 'required|numeric',
         ]);
-        $invoice = invoice::find($request->invoice_id);
-        $store = store::find($invoice->store_id);
-        $table = table::find($request->table_id);
-        if ($invoice->store_id == $table->store_id) {
-            if ($request->paidamount >= $invoice->f_discount) {
-                $table->status = 0;
-                $table->save();
-                $invoice->paid = $request->f_discount;
-                $invoice->save();
-                return 'Invoice Completed';
+
+        $positionApi = new positionApi();
+        $check = $positionApi->checkPositionRoute($request->store_id, Auth::id(), 'invoice_add');
+        if ($check) {
+            $invoice = invoice::find($request->invoice_id);
+            $store = store::find($invoice->store_id);
+            $table = table::find($request->table_id);
+            if ($invoice->store_id == $table->store_id) {
+                if ($request->paidamount >= $invoice->f_discount) {
+                    $table->status = 0;
+                    $table->save();
+                    $invoice->paid = $request->f_discount;
+                    $invoice->save();
+                    return 'Invoice Completed';
+                } else {
+                    return abort(419);
+                }
             } else {
-                return 'false';
+                return abort(419);
             }
         } else {
-            return 'fasle';
+            return abort(401);
         }
     }
